third_party/blink/web_tests/wpt_internal/scheduler/task_priority_change_error_on_recursive_call.html

<!doctype html>
<title>Scheduling API: Recursive TaskController.setPriority()</title>
<link rel="author" title="Scott Haseley" href="mailto:shaseley@chromium.org">
<link rel="help" href="https://github.com/WICG/scheduling-APIs">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
'use strict';

async_test(t => {
  let tc = new TaskController("user-visible");
  tc.signal.onprioritychange = t.step_func_done(() => {
    assert_equals(tc.signal.priority, "background");
    assert_throws_dom("NotAllowedError", () => { tc.setPriority("user-blocking"); });
  });
  tc.setPriority("background");
}, 'Test that TaskController.setPriority() throws an error if called recursively');

</script>
